fix(64bit): cast the remaining -Wsign-conversion sites (#15)
The non-rnd half of Phase 7, all representation-preserving casts that
match what the implicit conversion already did, so golden output is
byte-identical:

- qsort nmemb: the *_len() count (int) feeding qsort's size_t nmemb,
  cast (size_t) in bulk — gm.c (x5), perm.c (x4), report.c (x3),
  input.c (x2), use.c (x2), check.c, seed.c, swear.c.
- loop_known macro (olympia/loop.h): one (size_t) on its embedded
  qsort(ilist_len(kn)) clears all its expansion sites at once
  (gm.c x2, io.c, summary.c, report.c x2) — not separate edits.
- perm.c / sout.c: the fixed `spaces` buffer — (size_t)spaces_len in the
  perm.c subscript and my_malloc((size_t)spaces_len+1) in sout.c.
- mapgen/z.c: the my_malloc/my_realloc size-header + guard slots.
  G3-specific vs G2: G3's guard constants are 0xDEADBEEF/0xBABEFACE
  (exceed INT_MAX → unsigned-int-to-int), so each is an extra (int) cast
  beyond G2's lone (int)size — internal metadata, never emitted.
- mapgen/island.c: (size_t)(target_size + 1) on the island malloc count
  (G3-only, no G2 precedent — target_size is a non-negative count).

Golden gate YES (debug + asan-ubsan), asan/ubsan clean.

+++ mapgen/island.c
@@ -248,7 +248,7 @@ int main(int argc, char *argv[])
 			while (rnd(0, 99))
 				target_size++;
 	}
-	island = malloc((target_size + 1) * sizeof(*island));
+	island = malloc((size_t)(target_size + 1) * sizeof(*island));
 	if (!island) {
 		fprintf(stderr, "Unable to allocate memory for island data!\n");
 		exit(1);
